Updated for vdr-1.3.45. Fixed typos in audio tracks code (Thanks to Sebastian). Added Spanish translation (Thanks to Luis Palacios). Fixed compiling on AMD64 platform (Thanks to Mirco Günther and Jörg Bornkessel). Fixed date string in title bar (Thanks to Frank Krömmelbein for reporting this one). Added DXR3 logos from the enElchi-0.7.1 skin. Modified tags stripping of the auxiliary information field. Added VPS info for current event in channel info window.

+++ tools.h
@@ -12,7 +12,7 @@
 #include <vdr/tools.h>
 
 char *strcatrealloc(char *dest, const char *src);
-char *strncatrealloc(char *dest, const char *src, unsigned int len);
+char *strncatrealloc(char *dest, const char *src, size_t len);
 char *striptags(char *str);
 
 #endif // __SKINSOPPALUSIKKA_TOOLS_H
